Add lis3mdl spi bus and dataready support
Notably this does not disable GPIO interrupts in IRQ handler
like many of the other sensor drivers as when it was disabled
the trigger would often times be missed at higher sample rates.

This was likely caused by another interrupt of higher priority
interrupting our trigger function. This could be solved potentially
in the future by setting appropriate IRQ priorities.

This also adjusts the blerns board to not have conflicting pin
assignments, as the pin assignments for one of the UART ports
conflicted with what should be a SPI bus for the LIS3MDL chip.