rtio: Real-Time Stream API
A DMA friendly Stream API for zephyr.

Provides a way to define a driver which provides an output pipe of
blocks. Blocks are allocated from a specified memory pool and may be of
any allocatable size from the given memory pool. These blocks are easily
used for DMA transfers.

Some convience functions for storing and retrieving primitive types
in the blocks is provided.

Blocks are placed into the pipe based on a policy describing time or number of
entries since the block began being filled. If the block is full it is
meant to be immediately pushed into the pipe.

The use case in mind is high rate sensors with hardware FIFOs and SPI
interfaces where DMA transfers from these sensors would save a great
deal of CPU time avoiding the high rate of interrupts and small SPI
transfers that would otherwise occur. Other uses are certainly possible,
I imagine audio and video have similiar requirements.

Signed-off-by: Tom Burdick <thomas.burdick@gmail.com>