icm20649: TDK ICM20649 6 Axis Wide Range IMU
This adds an rtio_sensor implementation for the excellent TDK ICM20649
sensor with a large number of features.

Support for the hardware FIFO, power management, and various
filtering capabilities. Notably missing is the support for
this chip doing sensor fusion and acting as a i2c master for a
magnetometer.

Signed-off-by: Tom Burdick <thomas.burdick@gmail.com>