battery: Battery Device API
Battery fuel gauges and charge controllers are useful devices beyond
acting as a sensor. With a primary goal being to provide intelligent
notifications about battery status along with battery health.

While the Sensor API has historically been used for this I believe a
more specific Battery API that allows for setting battery policies
and providing a way to react to notifications when the battery goes
beyond the desired ranges is very useful.

A significant goal of this API is to support both embedded IoT type
battery charger/fuel gauge ic usage as well as the SBS specification.

The SBS specification can be seen as a superset of what a typical
charger and fuel gauge IC provides as they often have a small
microcontroller builtin to provide much of the notification and control.

Signed-off-by: Tom Burdick <thomas.burdick@intel.com>