soc/intel_adsp: hda test now uses WAIT_FOR
Avoids causing a wait with K_FOREVER which takes a lot longer to fail.

Doesn't use the _sync ipc commands anymore either as it seems sometimes
the interrupt doesn't fire (maybe printk spookiness as the other side
is in charge of pulling out of the window buf to print it).

Signed-off-by: Tom Burdick <thomas.burdick@intel.com>