logging: adsp_hda backend no flush locking
Use an atomic to flag to other calling contexts that
a log flush is already taking place and avoid waiting on
a spin lock when unnecessary. Without this flag the timer
callback would often times get stuck spinning for a long duration
as the hook function being called sends and waits on an IPC return.

Signed-off-by: Tom Burdick <thomas.burdick@intel.com>