sensor: RTIO API support for FIFO streaming
Provides the necessary APIs to stream a sensors FIFO using RTIO

Adds an API call to enable explicitly reading the fifo from the sensor
at any time. The thinking is that while the sensors gpio trigger might
be one source of triggering a fifo read, external sources should be
possible.

Signed-off-by: Tom Burdick <thomas.burdick@intel.com>