log: adsp_hda: Fixes immediate mode test
When calling the hook function outside of the spin lock the
timer may come along and subsequently call the hook function
again before it completes from the first call on another core.

This results in a deadlock waiting for IPC results is my belief.

Run the hook inside the safety of the spin lock and avoid that.

Signed-off-by: Tom Burdick <thomas.burdick@intel.com>