logging: adsp_hda: Fix many bugs
It turns out there were a few issues with HDA logging.

1. Concurrent attempts to notify with IPC outside of lock
2. The host falling behind and not accounting for that.
3. Not verifying free space before padding.
4. Cavstool not updating the hardware of its
   position (SPIB)

Greatly simplify the code in the log backend by
treating it like a little inline pipeline with steps. Each
step has a count of bytes it needs to work with. After each
step, the bytes for the next step are updated.

Only once all steps are completed are the bytes available to
use again.

Secondly just keep a never ending write_idx and use a power
of 2 mask for the ring.

Lastly wrap the entire thing in the spin lock.

This also updates cavstool to..

* correctly wait for the IPC done bit to be cleared by the
  hardware or warn if it doesn't happen.
* update the SPIB register to match the IO reader position
* ensure SPIB is enabled

Lastly it updates the log_immediate test to only use
1 core as using more than one core or having timer interrupts
disrupt the flow of logs causes some sort of horrible race
condition that I've yet to discover.

Signed-off-by: Tom Burdick <thomas.burdick@intel.com>