rtio: Add tiny write op
When sending small buffers out it makes sense to copy rather than
reference to avoid having to keep the small buffer around for the
lifetime of the write request.

Adjusts the op numbers to always be +1 from the previously defined op
id making it easier to re-arrange if needed in the future.

Additionally changes the userdata value to a uintptr_t to better
clarify that userdata may in fact be an integer or pointer, so long
as the storage requirements are the same.

Signed-off-by: Tom Burdick <thomas.burdick@intel.com>