icm42688: WIP to use SPI with RTIO
A work in progress to show a sensor driver that is an RTIO device
while internally using RTIO to perform sequences of SPI operations
to read the FIFO.

This enables a thread-free, lock-free FIFO reading chain driven entirely
by an interrupt.

The existing thread code to read the FIFO will remain ifdefry so a compare
and contrast scenario can be played out. What do latencies and throughputs
look like? What if we used data ready instead of the FIFO read but still
requested a batch read with the iodev?

There's many fun things we can try once SPI works.

Signed-off-by: Tom Burdick <thomas.burdick@intel.com>