spi: SAM use rtio when available even for blocking
This isn't really ideal as it requires a static queue to copy the
request into.

Instead I'm really thinking (another...) op code with the existing
spi parameters would be simpler. I could easily create a single sqe
on the stack...

Need to rework this a bit as it got a little ugly

Signed-off-by: Tom Burdick <thomas.burdick@intel.com>