sensing: Add a sensing bridge
The bridge is meant to take in multiple channel types and disperse the
data that comes in as if they are multiple sensors while taking in
arbitration requests and passing them to the single sensor driver.

Rather than a single phy_3d_sensor which is only able to wrap a single
channel of 3 axis data, a sensing bridge may wrap a single sensor device
and produce any number of split streams of data (accel, gyro, temp,
etc).

Signed-off-by: Tom Burdick <thomas.burdick@intel.com>