llext: Hello world for the shell_loader sample
Hello world from the tests is not longer what it once was, a simple
hello_world function that prints. The sample was here to let people
explore and learn about how llext and ELFs work.

Moving the sample extension code to the shell loader sample disconnects
the two and lets the sample remain simpler.

Signed-off-by: Tom Burdick <thomas.burdick@intel.com>