kernel: Optimize multiq for yielding
When yielding cooperative threads we can avoid some work knowing that we
are simply rotating the queue of threads at the same priority. This
requires changing the doubly linked list for the multiq to not include
the list management node as sys_dlist expects. So while the same
structures are being used here, the logic is customized.

Signed-off-by: Tom Burdick <thomas.burdick@intel.com>