llext: Dynamic (-fpic -shared) loading
In place of the fully static linking (relocatable) support -fpic -shared
built ELFs which generate global offset tables and procedure lookup
tables. These tables, which reside in .data, are the only things that
should need to be modified. When calling into functions the location of
the PLT resides in a dedicated register that is then used to address the
lookup tables where the final addresses of procedures and global
variables reside.

Effectively this will enable executing in place, or really anywhere you
like, with a lot less work and restrictions on our little embedded
linker.

Signed-off-by: Tom Burdick <thomas.burdick@intel.com>